physical_ai_k3s_llm/apps/central-hub/requirements.txt
flask
kubernetes
pyyaml
redis
requests